Add a description generator to create picnic description
The commit adds a `description-generator` module that generates picnic descriptions using OpenAI's API. The prompt includes the title, location, and date of the picnic. The prompt's response contains the description. The module is imported into the `/routes/picnics.js` file, the description is generated and then passed to the `createPicnic` method to store it in the database. In the `frontend`, the description is displayed along with the location, date, and the name of the picnic's host.